requirements.txt
python-telegram-bot==20.7
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
matplotlib==3.7.2
yfinance==0.2.28
python-dotenv==1.0.0